% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_box_metrics.R
\name{calculate_box_metrics}
\alias{calculate_box_metrics}
\title{Calculate Metrics for Multiple Boxes}
\usage{
calculate_box_metrics(boxes, genotypes)
}
\arguments{
\item{boxes}{A list of boxes (matrices) to analyze.}

\item{genotypes}{A vector of all possible genotypes.}
}
\value{
A list containing average and standard deviation of entropy, genotype counts, and Simpson's Diversity Index.
}
\description{
Computes aggregated metrics (entropy, unique genotypes, Simpson's Diversity Index, randomness percentage)
across multiple boxes.
}
\examples{
boxes <- list(
  list(box = matrix(c("cond1_G1_1", "cond1_G2_1", NA, NA), nrow = 2, byrow = TRUE)),
  list(box = matrix(c("cond1_G1_2", "cond1_G3_1", NA, NA), nrow = 2, byrow = TRUE))
)
genotypes <- c("G1", "G2", "G3")
calculate_box_metrics(boxes, genotypes)

}
