% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_box_metrics.R
\name{compute_box_metrics}
\alias{compute_box_metrics}
\title{Compute Box Metrics}
\usage{
compute_box_metrics(box)
}
\arguments{
\item{box}{A matrix representing the box.}
}
\value{
A list containing entropy, unique genotype count, and Simpson's Diversity Index.
}
\description{
Computes entropy, unique genotype count, and Simpson's Diversity Index for a box.
}
\examples{
box <- matrix(c("cond1_G1_1", "cond1_G2_1", "cond1_G1_2", NA), nrow = 2, ncol = 2)
compute_box_metrics(box)

}
