% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateBoxes.R
\name{generateBoxes}
\alias{generateBoxes}
\title{Generate Boxes}
\usage{
generateBoxes(
  samples,
  genotypes,
  reps,
  conditions,
  boxRows,
  boxCols,
  nIterations,
  nCores,
  availableBoxes = NULL
)
}
\arguments{
\item{samples}{Data frame with genotype and replicate information.}

\item{genotypes}{Column name for genotype data.}

\item{reps}{Column name for replicate data.}

\item{conditions}{Experimental conditions to consider.}

\item{boxRows}{Number of rows in each box.}

\item{boxCols}{Number of columns in each box.}

\item{nIterations}{Number of iterations to attempt box generation.}

\item{nCores}{Number of cores for parallel execution.}

\item{availableBoxes}{Optional; maximum number of boxes allowed.}
}
\value{
A list of generated boxes with conflict resolution applied.
}
\description{
Generates boxes of genotypic samples based on given conditions and parameters.
}
\examples{
samples <- data.frame(Genotype = c("G1", "G2"), Reps = c(3, 2))
generateBoxes(samples, "Genotype", "Reps", c("cond1"), 2, 2, 1, 1)

}
