% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGrids.R
\name{generateGrids}
\alias{generateGrids}
\title{Generate Boxes}
\usage{
generateGrids(
  samples,
  sample_labels,
  reps,
  conditions,
  gridRows,
  gridCols,
  nIterations,
  nCores,
  availableGrids = NULL
)
}
\arguments{
\item{samples}{Data frame with genotype and replicate information.}

\item{sample_labels}{Column name for genotype data.}

\item{reps}{Column name for replicate data.}

\item{conditions}{Experimental conditions to consider.}

\item{gridRows}{Number of rows in each box.}

\item{gridCols}{Number of columns in each box.}

\item{nIterations}{Number of iterations to attempt box generation.}

\item{nCores}{Number of cores for parallel execution.}

\item{availableGrids}{Optional; maximum number of grids allowed.}
}
\value{
A list of generated boxes with conflict resolution applied.
}
\description{
Generates boxes of genotypic samples based on given conditions and parameters.
}
\examples{
samples <- data.frame(Genotype = c("G1", "G2", "G3"), Reps = c(2, 2, 1))
conditions <- c("cond1", "cond2")
boxes <- generateGrids(samples, "Genotype", "Reps", conditions, 2, 2, 10, 1)

}
