% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_adjacent.R
\name{is_adjacent}
\alias{is_adjacent}
\title{Check for Adjacency Conflicts}
\usage{
is_adjacent(box, row, col, label)
}
\arguments{
\item{box}{A matrix representing the box.}

\item{row}{Row index of the cell to check.}

\item{col}{Column index of the cell to check.}

\item{label}{The genotype label in the cell.}
}
\value{
Logical value: \code{TRUE} if there is a conflict, \code{FALSE} otherwise.
}
\description{
Checks if a specific cell in a box matrix has adjacency conflicts
based on the given genotype label.
}
\examples{
box <- matrix(NA, nrow = 5, ncol = 5)
box[1, 1] <- "cond1_G1_1"
is_adjacent(box, 1, 2, "cond1_G1_2")

}
