% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_conflicts.R
\name{resolve_conflicts}
\alias{resolve_conflicts}
\title{Resolve Adjacency Conflicts}
\usage{
resolve_conflicts(box, max_iterations = 1000)
}
\arguments{
\item{box}{A matrix representing the box.}

\item{max_iterations}{Maximum number of iterations to attempt resolution.}
}
\value{
A matrix with resolved conflicts.
}
\description{
Resolves adjacency conflicts in a box by swapping labels between cells.
}
\examples{
box <- matrix(c(NA, "cond1_G1_1", "cond1_G1_2", NA), nrow = 2, ncol = 2)
resolve_conflicts(box, max_iterations = 100)

}
