% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreGrids.R
\name{scoreGrids}
\alias{scoreGrids}
\title{Score Boxes}
\usage{
scoreGrids(box_results)
}
\arguments{
\item{box_results}{A list of box generation results.}
}
\value{
A data frame with metrics for each experiment, including entropy, unique genotype counts,
and Simpson's Diversity Index.
}
\description{
Evaluates generated boxes and computes metrics for each experiment.
}
\examples{
box_results <- list(list(
  iteration = 1,
  boxes = list(
    cond1 = list(
      list(box = matrix(c("cond1_G1_1", "cond1_G2_1", NA, NA), nrow = 2, ncol = 2))
    )
  )
))
scoreGrids(box_results)

}
