% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_conflicts.R
\name{validate_conflicts}
\alias{validate_conflicts}
\title{Validate Adjacency Conflicts}
\usage{
validate_conflicts(box)
}
\arguments{
\item{box}{A matrix representing the box.}
}
\value{
Logical value: \code{TRUE} if no conflicts exist, \code{FALSE} otherwise.
}
\description{
Validates if any adjacency conflicts exist in a box.
}
\examples{
box <- matrix(c(NA, "cond1_G1_1", "cond1_G1_2", NA), nrow = 2, ncol = 2)
validate_conflicts(box)

}
